refactor: extract tessen route tracking to own file
