+++ packages/gatsby-theme-newrelic/gatsby/on-route-update.js
@@ -1,45 +1,10 @@
 import { withDefaults } from '../src/utils/defaultOptions';
-import createTessen from '../src/utils/createTessen';
-import warning from 'warning';
+import trackViaTessen from '../src/utils/page-tracking/tessen';
 
-const warnAboutNoop = (pageView) => {
-  warning(
-    pageView,
-    "You have enabled page view tracking, but do not have page view tracking configured. This route change has not been tracked. Please configure the 'tessen.pageView' option in gatsby-config.js"
-  );
-
-  warning(
-    pageView.name && pageView.category,
-    "You have enabled page view tracking, but page view tracking is misconfigured. This route change has not been tracked. Please configure the 'tessen.pageView.name' and 'tessen.pageView.category' options in gatsby-config.js"
-  );
-};
-
-const canSendPageView = (pageView) =>
-  pageView && pageView.name && pageView.category;
-
-const onRouteUpdate = ({ location }, themeOptions) => {
+const onRouteUpdate = ({ location, prevLocation }, themeOptions) => {
   const { env, tessen: tessenConfig } = withDefaults(themeOptions);
 
-  if (!tessenConfig || !tessenConfig.trackPageViews) {
-    return;
-  }
-
-  const { pageView } = tessenConfig;
-  const tessen = createTessen(tessenConfig);
-
-  if (!canSendPageView(pageView)) {
-    return warnAboutNoop(pageView);
-  }
-
-  const { name, category, getProperties } = pageView;
-
-  requestAnimationFrame(() => {
-    tessen.page(
-      name,
-      category,
-      getProperties && getProperties({ location, env })
-    );
-  });
+  trackViaTessen({ location, prevLocation }, tessenConfig, env);
 };
 
 export default onRouteUpdate;
